requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
curl==0.0.1
Django==3.1.5
django-cors-middleware==1.5.0
django-environ==0.4.5
django-extensions==3.1.0
django-haystack==3.0
django-oauth-toolkit==1.3.3
django-shortcuts==1.6
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
pytz==2020.5
requests==2.25.1
South==1.0.2
sqlparse==0.4.1
urllib3==1.26.3
